# apps/taptrade-platform/go-platform/services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Taya NA Predict API
  version: "1.0.0"
  description: >
    TapTrade prediction-market gateway API for non-redeemable gameplay points.
    Binary YES/NO outcomes use point-cents subunits on a 0-100 implied
    probability scale; settled winning shares return 100 point-cents per share.
    Points are gameplay-only units with no external value path.

    Coverage: this spec documents the core public, trading, portfolio,
    responsible-play, reward, loyalty, social, leaderboard, bot, auth proxy,
    admin account-review, admin discovery taxonomy, admin market operations,
    and admin market-lifecycle surfaces verified against the live handlers and
    DTOs as of 2026-06. CMS and remaining compatibility surfaces are
    intentionally out of scope here and tracked for documentation in
    IMPROVEMENT_PLAN.md P2-05 / G-04. The previous inherited spec is archived
    outside this launch document.
servers:
  - url: http://localhost:18080
    description: Local development
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly session cookie issued by the auth service. Mutating requests also require the X-CSRF-Token header (double-submit cookie).
    BotApiKey:
      type: http
      scheme: bearer
      description: "Bot/partner API key: Authorization: Bearer tna_<prefix>_<secret>. Scopes: read, trade."
  schemas:
...
    BonusClaimRequest:
      type: object
      properties:
        campaign_id: { type: integer, format: int64 }
        trigger_reference: { type: string, example: "daily-check-in:2026-06-26" }
      required: [campaign_id, trigger_reference]
    PlayerBonus:
      type: object
      properties:
        unit: { type: string, enum: [PTS] }
        bonus_id: { type: integer, format: int64 }
        bonusId: { type: integer, format: int64 }
        campaign_name: { type: string }
        campaignName: { type: string }
        bonus_type: { type: string, enum: [signup_bonus, custom, point_grant, point_match] }
        bonusType: { type: string, enum: [signup_bonus, custom, point_grant, point_match] }
        status: { type: string }
        expires_at: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        granted_at: { type: string, format: date-time }
        grantedAt: { type: string, format: date-time }
        grantedPoints: { type: integer, format: int64 }
        remainingPoints: { type: integer, format: int64 }
        playRequiredPoints: { type: integer, format: int64 }
        playCompletedPoints: { type: integer, format: int64 }
        playProgressPct: { type: number, format: double }
    PlayerBonusListResponse:
      type: object
      properties:
        bonuses:
          type: array
          items: { $ref: "#/components/schemas/PlayerBonus" }
...
    AdminCampaign:
      type: object
      properties:
        unit: { type: string, enum: [PTS] }
        id: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string }
        campaignType: { type: string, enum: [signup_bonus, custom, point_grant, point_match] }
        status: { type: string }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        budgetPoints: { type: integer, format: int64, nullable: true }
        spentPoints: { type: integer, format: int64 }
        maxClaims: { type: integer, nullable: true }
        claimCount: { type: integer }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AdminCampaignListResponse:
      type: object
      properties:
        campaigns:
          type: array
          items: { $ref: "#/components/schemas/AdminCampaign" }
...
    AdminCampaignRuleInput:
      type: object
      properties:
        rule_type: { type: string, enum: [eligibility, trigger, reward, play] }
        point_rule_config:
          type: object
          additionalProperties: true
          description: Point-native campaign rule config accepted by launch admin clients; play contribution caps use max_play_contribution_points.
      required: [rule_type, point_rule_config]
    AdminCampaignCreateRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        campaign_type: { type: string, enum: [signup_bonus, custom, point_grant, point_match] }
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        budget_points: { type: integer, format: int64, nullable: true }
        max_claims: { type: integer, nullable: true }
        rules:
          type: array
          items: { $ref: "#/components/schemas/AdminCampaignRuleInput" }
      required: [name, campaign_type, start_at, end_at]
...
    AdminBonusGrantRequest:
      type: object
      properties:
        user_id: { type: string }
        campaign_id: { type: integer, format: int64 }
        override_points: { type: integer, format: int64, nullable: true }
        reason: { type: string }
      required: [user_id, campaign_id, reason]
...
paths:
...
  /api/v1/bonuses/active:
    get:
      summary: List active point-play bonuses for the caller
      description: Returns only PTS unit fields and point-play progress aliases for the authenticated session user.
      security: [{ SessionCookie: [] }]
      responses:
        "200":
          description: Active bonus grants
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayerBonusListResponse" }
        "401": { description: Authentication required }
  /api/v1/bonuses/claim:
    post:
      summary: Claim an eligible point-play bonus campaign
      description: Creates a player bonus grant for the authenticated session user and returns point-native fields only.
      security: [{ SessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BonusClaimRequest" }
      responses:
        "201":
          description: Claimed point-play bonus
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayerBonus" }
        "401": { description: Authentication required }
        "403": { description: Campaign is unavailable for this session user }
        "409": { description: Campaign was already claimed }
...
  /api/v1/admin/campaigns:
    get:
      summary: Admin list of point-play campaigns
      description: Returns campaign rows with PTS unit, budgetPoints, and spentPoints aliases only.
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: Campaign rows
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminCampaignListResponse" }
        "403": { description: Admin role required }
    post:
      summary: Create a point-play campaign
      description: Creates campaign metadata and returns the point-native campaign response contract.
      security: [{ SessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminCampaignCreateRequest" }
      responses:
        "201":
          description: Created point-play campaign
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminCampaign" }
        "400": { description: Invalid campaign request }
        "403": { description: Admin role required }
...
  /api/v1/admin/bonuses:
    get:
      summary: Admin list of player point-play bonus grants
      description: Returns point-native player bonus contracts for office review.
      security: [{ SessionCookie: [] }]
      parameters:
        - { name: user_id, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: Player bonus grants
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayerBonusListResponse" }
        "403": { description: Admin role required }
  /api/v1/admin/bonuses/grant:
    post:
      summary: Admin grant of a point-play bonus
      description: Grants a player bonus and returns the same PTS response contract as player-facing bonus reads.
      security: [{ SessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminBonusGrantRequest" }
      responses:
        "201":
          description: Granted player bonus
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayerBonus" }
        "400": { description: Invalid grant request }
        "403": { description: Admin role required }
...